Fix root cause of type promotion from 32bit to 64bit float

+++ pytorch_forecasting/data/encoders.py
@@ -355,10 +355,10 @@ class TorchNormalizer(BaseEstimator, TransformerMixin):
                 self.scale_ = torch.std(y_scale, dim=-1) + self.eps
             elif isinstance(y_center, np.ndarray):
                 self.center_ = np.mean(y_center, axis=-1)
-                self.scale_ = np.std(y_scale, axis=-1) + self.eps
+                self.scale_ = (np.std(y_scale, axis=-1) + self.eps).astype(y_scale.dtype)
             else:
                 self.center_ = np.mean(y_center)
-                self.scale_ = np.std(y_scale) + self.eps
+                self.scale_ = (np.std(y_scale) + self.eps).astype(y_scale.dtype)
 
         elif self.method == "robust":
             if isinstance(y_center, torch.Tensor):
